Match the remaining code

+++ include/game/objects/cutscene/library_puzzle.h
@@ -61,7 +61,7 @@ void libraryPuzzle_puzzle_success(libraryPuzzle* self);
 void libraryPuzzle_destroy(libraryPuzzle* self);
 void libraryPuzzle_printSelectedOptions(u16* text, u16 selected_options_IDs);
 s32 libraryPuzzle_selectNextOption(
-    u32* highlighted_option, s16* selection_delay_timer, u32* selected_options_IDs
+    s32* highlighted_option, u16* selection_delay_timer, u16* selected_options_IDs
 );
 
 typedef enum libraryPuzzle_func_id {
